added warning for #version mismatch in shadercode
I spent literal days trying to figure out why I couldn't use #version 300 shader code, and it's because I was only modifying the fragment shader. 

Because this code only looks for the version directive in the vertex shader, it will cause an error if you have a version directive in only the fragment shader. Since detection of #version directives is completely undocumented, there is no way to know that the solution is to simply add a #version directive to the vertex shader. 

I would suggest that the whole if statement be re-worked so that it actually properly handles #version directives while also properly injecting it's own code, but I don't fully understand all of what is going on here so I'm not confident in doing it. As it stands, this will at least get confused devs pointed in the right direction.